[TASK] epics controller to common controller
